requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=9.0.0

# Computer Vision
opencv-python>=4.5.0
face-recognition>=1.3.0
mtcnn>=0.1.1

# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.5
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Monitoring & Metrics
prometheus-client>=0.17.0
prometheus-fastapi-instrumentator>=6.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=0.19.0

# Utilities
aiofiles>=23.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.24.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
pre-commit>=3.3.0

# Additional Dependencies
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
